First stab at PHP functions calling Python.
At the moment, this is intentionally restrictive: only Python objects wrapped as
W_EmbeddedPyFunc can be called. In other words, with this scheme, Python objects
which define a __call__ method can't be called from PHP. We can loosen this
restriction as and when we work out what we want to do in such cases.