Add plumbing to detect when objects wrap a different VM's objects.
Before this patch, there was no sensible way to ask a PHP object "do you wrap a
Python object" (or vice versa). One could only do so via torturous isinstance
checks, which are likely to be brittle, assuming we add more fine-grained
wrappers in the future.

[While here, we also need a way to get the PHP interpreter from wrapped objects.
We will probably need a similar mechanism in the opposite direction in the
future.]

Note: there is a corresponding patch in PyPy.