Experimentally rename W_PyBridgeProxy to W_PyProxyGeneric.
It's odd to have Bridge in that classname and no other (and not particularly
helpful). Ideally everything in php_wrappers would be of the form W_PyProxy<x>,
but I won't do anything else until Edd's merged other branches in.

While here, rename the magic__ methods to something more helpful.