Merge Laurie's immut cache change from pypy_bridge branch.
Original commit message:

Alter the immutable cache class so it returns None instead of raising KeyError.

This allows code that is searching for something that it doesn't hugely expect
to find to become much more JIT friendly. On Richards, this leads to a 1/3 speed
increase. This does touch a lot of the code-base, so there is a small chance
that it negatively effects something.

Conflicts:
	hippy/interpreter.py
	hippy/module/pypy_bridge/scopes.py