Allow Python to reference dynamically created PHP variables.
Previously this only worked if there was a static reference to a dynamically
created variable (which is not hugely useful).

Whether this is what we want or not is a different question. There is a big,
subtle, question to be asked (and, one would like to think, answered) about the
right combination of static and dynamic scoping.