Do not read e.__class__ as it causes -Ojit translation to fail
(this was broken since PR #54)

The reason for the failure is deep RPython nonsense:
The JIT defines a JitException which subclasses Exception and is only seen
during the JIT translation phase. This triggers a call
to rpython.annotator.classdef.ClassDef.see_new_subclass, which attempts
to reflow main() because it accesses <Exception>.__class__,
which fails because main() has already been rtyped at
this point.