Move W_PyException to the right place.
The naming of py_wrappers.py and php_wrappers.py is really confusing.

Is a py_wrapper a PHP object wrapping a python object or a Python object
that wraps a PHP object. The name could mean either.

As it stands, py_wrappers means the outer wrapper is for python, thus
wrapping a PHP object, but I wish we could find better naming conventions.