Add only_ref_arrays option to getattr_ref.
When we're looking up an attribute, we're in a sticky situation if we look up an
array and then mutate it. Calling getattr() would (conceptually) create a copy
of the array, so Python's mutations wouldn't show up in PHP. This code was thus
altered a while back to use getattr_ref, which returns a reference to a thingy
rather than the thingy itself. Mutation then happens on the reference, and
everything looks sensible to both Python and PHP. However, getting a reference
to any PHP thingy is slow and, pointless, since most items are immutable (i.e.
there's no point getting a reference to a PHP int, because we can't mutate the
int).

This patch adds a knob to getattr_ref called only_ref_arrays which only adds a
new reference if the attribute we're looking for happens to be an array. This
gets us the semantics we want without a performance hit for the non-array case.