Overhaul the way PHP sees Python classes.
Namely implement PyCalssAdapter and PyCalssAdapterClass.

Fixes a xfail test, but sadly slows down richards and deltablue a
little. After much debugging, it turns out to be a call to a vref
self.topframeref() in hippy.interpreter.Interpreter which cannot
be traced.

Have added an XXX pointing out the troublesome line. Treating this as a
separate issue and will work on this is a separate branch.

This is a combination of 21 commits.