Bring call_py_func into shape.
Re-enable the ability to call static methods.
Allow calling of arbitrary callable instances.
Re-factor.

Made sure the function is still tracable. Benchmarks show a small slowdown,
but performance is still generally good. It is still much faster than
before this function was traced for example.

Probably room for more optimisation, but we have bigger fish to fry at
the moment.

(This is a combination of 14 commits.)